+++ source/Public/Restart-SqlDscRSService.ps1
@@ -132,7 +132,7 @@ function Restart-SqlDscRSService
         }
         catch
         {
-            $errorMessage = $script:localizedData.Restart_SqlDscRSService_ServiceNotFound -f $targetServiceName, $_.Exception.Message
+            $errorMessage = $script:localizedData.Restart_SqlDscRSService_ServiceNotFound -f $targetServiceName
 
             Write-Error -Message $errorMessage -Category ObjectNotFound -ErrorId 'RSRSRS0001' -TargetObject $targetServiceName